requirements.txt
matplotlib==3.9.0
nltk==3.8.1
numpy==2.0.0
openai==1.35.3
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
spacy==3.7.5
streamlit==1.36.0
textblob==0.18.0.post0
tensorflow==2.16.1
torch==2.3.1
transformers==4.41.2
vaderSentiment==3.3.2
wordcloud==1.9.3
xgboost==2.0.3
